+++ tests/QueryType/Select/Result/Stats/FacetValueTest.php
@@ -24,11 +24,21 @@ class FacetValueTest extends TestCase
             'max' => 3.5,
             'sum' => 3.333,
             'count' => -1,
-            'missing' => 'dummyMissing',
+            'missing' => 0,
             'sumOfSquares' => 17.23,
-            'mean' => 'dummyMean',
+            'mean' => 3.14,
             'stddev' => 1.11,
-            'facets' => ['dummyFacets'],
+            'percentiles' => [
+                '50.0' => 3.14,
+                '90.0' => 42.0,
+            ],
+            'distinctValues' => [
+                'dummy1',
+                'dummy2',
+            ],
+            'countDistinct' => 3,
+            'cardinality' => 2,
+            'dummy' => 'test',
         ];
 
         $this->result = new FacetValue($this->value, $this->stats);
@@ -79,8 +89,41 @@ class FacetValueTest extends TestCase
         $this->assertSame($this->stats['stddev'], $this->result->getStddev());
     }
 
+    public function testGetPercentiles()
+    {
+        $this->assertSame($this->stats['percentiles'], $this->result->getPercentiles());
+    }
+
+    public function testGetDistinctValues()
+    {
+        $this->assertSame($this->stats['distinctValues'], $this->result->getDistinctValues());
+    }
+
+    public function testGetCountDistinct()
+    {
+        $this->assertSame($this->stats['countDistinct'], $this->result->getCountDistinct());
+    }
+
+    public function testGetCardinality()
+    {
+        $this->assertSame($this->stats['cardinality'], $this->result->getCardinality());
+    }
+
+    /**
+     * @deprecated Will be removed in Solarium 7.
+     */
     public function testGetFacets()
     {
-        $this->assertSame($this->stats['facets'], $this->result->getFacets());
+        $this->assertNull($this->result->getFacets());
+    }
+
+    public function testGetStatValue()
+    {
+        $this->assertSame($this->stats['dummy'], $this->result->getStatValue('dummy'));
+    }
+
+    public function testGetStatValueUnknown()
+    {
+        $this->assertNull($this->result->getStatValue('unknown'));
     }
 }